pyproject.toml
[project]
name = "moe-pretraining"
version = "0.1.0"
description = "Standard vs Global MoE pretraining experiments"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.0",
    "accelerate>=0.30.0",
    "transformers>=4.40.0",
    "datasets>=2.19.0",
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "wandb>=0.17.0",
    "einops>=0.7.0",
    "tokenizers>=0.19.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
